+++ libraries/engage/components/View/components/Content/style.js
@@ -1,9 +1,9 @@
 import { css } from 'glamor';
-import { useScrollContainer } from '@shopgate/engage/core';
+import { useScrollContainer, isIOs } from '@shopgate/engage/core/helpers';
 import { responsiveMediaQuery } from '@shopgate/engage/styles';
 import { IS_PAGE_PREVIEW_ACTIVE } from '@shopgate/engage/page/constants';
 
-export default css({
+export const container = css({
   display: 'flex',
   flexDirection: 'column',
   width: '100vw',
@@ -23,5 +23,19 @@ export default css({
   }),
   [responsiveMediaQuery('>xs', { webOnly: true })]: {
     width: 'var(--page-content-width)',
+  },
+});
+
+export const containerInner = css({
+  ...isIOs && useScrollContainer() ? {
+    // Make the scroll container content a bit higher than the actual scroll container to
+    // get a rubber band effect in all situations
+    minHeight: 'calc(100% + var(--extra-ios-scroll-space, 0px))',
+  } : {},
+  ':after': {
+    content: "''",
+    display: 'block',
+    pointerEvents: 'none',
+    paddingBottom: 'calc(var(--page-content-offset-bottom) + var(--keyboard-height))',
   },
 });