+++ libraries/engage/locations/actions/sendDefaultLocationCode.js
@@ -1,3 +1,8 @@
+import {
+  sendSetDefaultLocationCodeRequest,
+  sendSetDefaultLocationCodeSuccess,
+  sendSetDefaultLocationCodeError,
+} from '@shopgate/engage/locations/action-creators';
 import { PipelineRequest } from '../../core';
 
 /**
@@ -6,8 +11,22 @@ import { PipelineRequest } from '../../core';
  * @param {string} locationCode The location code to send to backend
  * @returns {Function} redux thunk
  */
-const sendDefaultLocationCode = locationCode => () => new PipelineRequest('shopgate.user.setDefaultLocationCode')
-  .setInput({ locationCode })
-  .dispatch();
+const sendDefaultLocationCode = locationCode => (dispatch) => {
+  dispatch(sendSetDefaultLocationCodeRequest(locationCode));
+
+  const request = new PipelineRequest('shopgate.user.setDefaultLocationCode')
+    .setInput({ locationCode })
+    .dispatch();
+
+  request
+    .then(() => {
+      dispatch(sendSetDefaultLocationCodeSuccess(locationCode));
+    })
+    .catch((error) => {
+      dispatch(sendSetDefaultLocationCodeError(locationCode, error.code));
+    });
+
+  return request;
+};
 
 export default sendDefaultLocationCode;