+++ libraries/engage/product/components/index.js
@@ -28,3 +28,5 @@ export * from './RelationsSlider';
 export * from './Swatch';
 export * from './Swatches';
 export * from './UnitQuantityPicker';
+export { default as FilterBar } from './FilterBar';
+export { default as ProductFilters } from './ProductFilters';